Updating collision avoidance

+++ Assets/Scripts/Capybara/CapybaraInfo.cs
@@ -6,7 +6,7 @@ public class CapybaraInfo : MonoBehaviour
 {
     public string capyName;
     public float happiness = 75, hunger = 75, comfort = 75, fun = 75;
-    public int Frustration { get; set; }
+    public float Frustration { get; set; }
     public float TimeSpent { get; set; }
     public bool HungerFilling { get; set; } = false;
     public bool ComfortFilling { get; set; } = false;
